+++ packages/two_dimensional_scrollables/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.9
+
+* Fixes TableSpan borders being flipped when one or both axis directions are reversed.
+
 ## 0.3.8
 
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.